requirements.txt
psycopg2-binary==2.9.1
numpy==1.21.2
pandas==1.3.3
logger==1.4
sqlalchemy==1.4.25
pyodbc==4.0.32
configparser==5.0.2
importlib==1.0.4
requests==2.26.0
rpy2==3.4.5